Refactoring predict.py file to split it to other files
