+++ landingai/predict/__init__.py
@@ -1,0 +1,4 @@
+from .edge import EdgePredictor
+from .cloud import Predictor
+from .ocr import OcrPredictor
+from .snowflake import SnowflakeNativeAppPredictor
